+++ app/resources/numismatics/issues/numismatics/issue_wayfinder.rb
@@ -16,7 +16,11 @@ module Numismatics
     end
 
     def issues_count
-      @issues = query_service.custom_queries.count_all_of_model(model: Numismatics::Issue)
+      @issues_count = query_service.custom_queries.count_all_of_model(model: Numismatics::Issue)
+    end
+
+    def issues
+      @issues = query_service.find_all_of_model(model: Numismatics::Issue)
     end
 
     def members_with_parents